mod_aes67: cleanup session elements in the channel_hangup
Move the cleanup of gstreamer elements associated to a session from the
channel_kill to channel_hangup so that they can be torn down asynchronously
when the HANGUP events are dispatched.

This avoids blocking of the channel_kill which is called while terminating a
session using uuid_kill or hupall commands from the fs_cli

The issue is mostly seen when there are multiple sessions being originated
and if we issue command to terminate all sessions concurrently.